pyproject.toml
[project]
name = "tayend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi-crons>=2.0.1",
    "fastapi[standard]>=0.116.1",
    "portia-sdk-python[all]>=0.6.2",
    "sse-starlette>=3.0.2",
]
